# PERF: release gil for ewma_time