--- a/pandas/_libs/window/aggregations.pyx
+++ b/pandas/_libs/window/aggregations.pyx
@@ -1368,7 +1368,7 @@ def roll_weighted_var(float64_t[:] values, float64_t[:] weights,
 # ----------------------------------------------------------------------
 # Exponentially weighted moving average
 
-def ewma_time(float64_t[:] vals, int minp, ndarray[int64_t] times,
+def ewma_time(const float64_t[:] vals, int minp, ndarray[int64_t] times,
               int64_t halflife):
     """
     Compute exponentially-weighted moving average using halflife and time