--- a/pandas/_libs/window/aggregations.pyx
+++ b/pandas/_libs/window/aggregations.pyx
@@ -1417,7 +1417,7 @@ def ewma_time(const float64_t[:] vals, int minp, ndarray[int64_t] times,
     cdef:
         Py_ssize_t i, j, num_not_nan = 0, N = len(vals)
         bint is_not_nan
-        float64_t last_result, weights_dot, weights_sum, weight
+        float64_t last_result, weights_dot, weights_sum, weight, halflife_flot
         float64_t[:] times_float
         float64_t[:] observations = np.zeros(N, dtype=float)
         float64_t[:] times_masked = np.zeros(N, dtype=float)
@@ -1426,6 +1426,7 @@ def ewma_time(const float64_t[:] vals, int minp, ndarray[int64_t] times,
     if N == 0:
         return output
 
+    halflife_flot = <float>halflife
     times_float = times.astype(float)
     last_result = vals[0]
 
@@ -1440,7 +1441,7 @@ def ewma_time(const float64_t[:] vals, int minp, ndarray[int64_t] times,
                 weights_sum = 0
                 weights_dot = 0
                 for j in range(num_not_nan):
-                        weight = 0.5 ** ((times_float[i] - times_masked[j]) / halflife)
+                        weight = 0.5 ** ((times_float[i] - times_masked[j]) / halflife_flot)
                         weights_sum += weight
                         weights_dot += weight * observations[j]