The performance changes for `rolling.EWMMethods` are not stable. I've seen those and no changes in separate runs.
```
+     1.35±0.02ms       2.17±0.5ms     1.61  rolling.EWMMethods.time_ewm('Series', 1000, 'float', 'std')
```
```
-      2.58±0.2ms      1.36±0.01ms     0.53  rolling.EWMMethods.time_ewm('Series', 10, 'float', 'std')
```
```
+     1.39±0.01ms       2.17±0.5ms     1.56  rolling.EWMMethods.time_ewm('DataFrame', 10, 'float', 'std')
-      1.71±0.3ms      1.35±0.01ms     0.79  rolling.EWMMethods.time_ewm('Series', 10, 'float', 'std')
```