@jbrockmendel 2020-10-28
did you ever get consistent results from the asvs?