@fangchenli 2020-10-29
> did you ever get consistent results from the asvs?

No, and I have no idea why...